Properly cast (time / period) before assigning to the counter

+++ Authenticator/Classes/OTPToken+Generation.m
@@ -68,7 +68,7 @@ static NSUInteger kPinModTable[] = {
         if (self.isInKeychain)
             [self saveToKeychain];
     } else if (self.type == OTPTokenTypeTimer) {
-        self.counter = ([NSDate date].timeIntervalSince1970 / self.period);
+        self.counter = (uint64_t)([NSDate date].timeIntervalSince1970 / self.period);
     }
 }
